# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Media Circle API
  description: Rest api for medialister
  version: 1.0.0
  contact:
    name: Media Circle
servers:
  - url: http://localhost:8080/api/v1
    description: Development server

paths:
  /liste/{id}: 
    get:
      parameters:
        - name: id
          in: path
          description: Hvilken liste skal hentes
          schema:
            type: integer
      description: Henter liste
      operationId: getListe
      tags:
        - liste
      responses:
        '200':
          description: objekter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Liste'
  /liste/test/{id}: 
    get:
      parameters:
        - name: id
          in: path
          description: Hvilken liste skal hentes
          schema:
            type: integer
      description: Henter liste
      operationId: getListeTest
      tags:
        - liste
      responses:
        '200':
          description: objekter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Liste'
  /anmeldelse/slim/{brukerId}:
    get:
      parameters:
        - name: brukerId
          in: path
          description: Hvilken bruker det skal hentes fra
          schema:
            type: string
      description: Henter alle slim anmeldelser for en bruker
      operationId: getAlleAnmeldelserForBruker
      tags:
        - anmeldelse
      responses:
        '200':
          description: list of anmeldelseSlim
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnmeldelseSlim'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
  /albums:
    get:
      summary: henter alle album
      description: henter en liste med alle album
      operationId: getAlbums
      tags:
        - albums
      responses:
        '200':
          description: list of albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
    
    post:
      summary: nytt album
      description: legger til et nytt album i databasen
      operationId: createAlbum
      tags:
        - albums
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlbumRequest'
      responses:
        '201':
          description: 200 ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
components:
  schemas:
    Album:
      type: object
      required:
        - id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          minLength: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AnmeldelseSlim:
      type: object
      required:
        - id
        - navnPaaMedie
        - antallStjerner
      properties:
        id:
          type: integer
          format: int64
        antallStjerner:
          type: integer
          format: int64
        navnPaaMedie:
          type: string
          minLength: 1
    CreateAlbumRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
    Liste:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
    Response:
      type: object
      required:
        - message
        - status
      properties:
        message:
          type: string
        status:
          type: integer
          example: 200