# backend/generated/openapi_types.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package generated

import (
	"time"
)

// Album defines model for Album.
type Album struct {
	CreatedAt time.Time `json:"created_at"`
	Id        int64     `json:"id"`
	Name      string    `json:"name"`
	UpdatedAt time.Time `json:"updated_at"`
}

// AnmeldelseSlim defines model for AnmeldelseSlim.
type AnmeldelseSlim struct {
	AntallStjerner int64  `json:"antallStjerner"`
	Id             int64  `json:"id"`
	NavnPaaMedie   string `json:"navnPaaMedie"`
}

// CreateAlbumRequest defines model for CreateAlbumRequest.
type CreateAlbumRequest struct {
	Name string `json:"name"`
}

// Liste defines model for Liste.
type Liste struct {
	Name string `json:"name"`
}

// Response defines model for Response.
type Response struct {
	Message string `json:"message"`
	Status  int    `json:"status"`
}

// CreateAlbumJSONRequestBody defines body for CreateAlbum for application/json ContentType.
type CreateAlbumJSONRequestBody = CreateAlbumRequest